Deprecate getRequestHostname and replace it with getRequestHost.
Host names are a feature of the transport over which HTTP runs. As of
HTTP 1.1, however, the host of a request is a distinct,
application-protocol-level property specified by the Host header.
HTTP 1.0 did not require or even explicitly describe the host header,
and an HTTP 1.0 request that omits it is conventionally understood to
have a host derived from the transport's hostname.

HTTP 1.1, however, ​requires a Host header; requests that omit it
should be rejected with a 400 status code, but that is outside the
scope of this commit.

To make the primacy of an HTTP's host (over its transport hostname)
clearer, Request.getRequestHostname should be deprecated and replaced
with Request.getRequestHost.  This should have the same behavior,
except that it should raise an exception for HTTP 1.0 requests that
both lack a host header and run over a transport that does not have a
host name (e.g., a UNIX domain socket.)